Remove launcher icon from Cell Broadcast app.
Update the Cell Broadcast manifest to remove the launcher icon
for the main activity. It will now be accessible via a new menu
item in the MMS app, as well as a new menu item in Settings to
jump directly to the Cell Broadcast settings activity, both of
which are enabled by setting "config_cellBroadcastAppLinks" to
true in a device config overlay.

Bug: 6709985
Change-Id: I37d2894ed68e8856a112780f27f03864a57162db